# packages/functions/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Alexandria API
  description: API for managing libraries and books
  version: 1.0.0
  contact:
    name: Alexandria
servers:
  - url: /v1
    description: API v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Detection
...
    ItemType:
      type: integer
      enum: [0, 1]
      description: "Item type (0 = Book, 1 = Video)"
    GetItemResponseBase:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/ItemType"
        title:
          type: string
        picture:
          type: string
          nullable: true
          description: "Base64 encoded picture"
        pictureUrl:
          type: string
          nullable: true
        libraryId:
          type: string
          nullable: true
        libraryName:
          type: string
          nullable: true
        ownerId:
          type: string
        lentTo:
          type: string
          nullable: true
          description: "Name of person the item is lent to"
        collectionId:
          type: string
          nullable: true
          description: "Collection ID"
        collectionName:
          type: string
          nullable: true
          description: "Collection name (denormalized)"
        order:
          type: integer
          nullable: true
          description: "Order within collection (1-1000)"
    GetBookResponse:
      allOf:
        - $ref: "#/components/schemas/GetItemResponseBase"
        - type: object
          properties:
            authors:
              type: array
              items:
                type: string
            summary:
              type: string
            isbn:
              type: string
...
    GetLibrariesContentResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GetBookResponse"
        nextToken:
          type: string
          description: "Pagination token for next page"
    CreateBookRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        summary:
          type: string
          maxLength: 4000
        authors:
          type: array
          items:
            type: string
        isbn:
          type: string
        pictureUrl:
          type: string
          nullable: true
        collectionId:
          type: string
          nullable: true
          description: "Collection ID to add the book to"
        order:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1000
      required:
        - title
    CreateBookResponse:
      type: object
      properties:
        id:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        summary:
          type: string
          maxLength: 4000
        authors:
          type: array
          items:
            type: string
        isbn:
          type: string
        pictureUrl:
          type: string
          nullable: true
        collectionId:
          type: string
          nullable: true
          description: "Collection ID (null to remove from collection)"
        order:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1000
        updatePicture:
          type: boolean
          nullable: true
          description: "If true, fetch and update the picture from pictureUrl"
      required:
        - title
    CreateVideoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        summary:
          type: string
          maxLength: 4000
        directors:
          type: array
          items:
            type: string
            maxLength: 100
        cast:
          type: array
          items:
            type: string
        releaseYear:
          type: integer
          nullable: true
          minimum: 1800
          maximum: 2100
        duration:
          type: integer
          nullable: true
          minimum: 0
          maximum: 1000
          description: "Duration in minutes"
        tmdbId:
          type: string
          nullable: true
        pictureUrl:
          type: string
          nullable: true
        collectionId:
          type: string
          nullable: true
          description: "Collection ID to add the video to"
        order:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1000
      required:
        - title
    CreateVideoResponse:
      type: object
      properties:
        id:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
    UpdateVideoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        summary:
          type: string
          maxLength: 4000
        directors:
          type: array
          items:
            type: string
            maxLength: 100
        cast:
          type: array
          items:
            type: string
        releaseYear:
          type: integer
          nullable: true
          minimum: 1800
          maximum: 2100
        duration:
          type: integer
          nullable: true
          minimum: 0
          maximum: 1000
          description: "Duration in minutes"
        tmdbId:
          type: string
          nullable: true
        pictureUrl:
          type: string
          nullable: true
        collectionId:
          type: string
          nullable: true
          description: "Collection ID (null to remove from collection)"
        order:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1000
        updatePicture:
          type: boolean
          nullable: true
          description: "If true, fetch and update the picture from pictureUrl"
      required:
        - title

    # Item History
    ItemEventType:
      type: string
      enum:
        - LENT
        - RETURNED
    ItemHistoryEntryRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/ItemEventType"
        event:
          type: string
          maxLength: 50
          description: "Name of person (for lend/return)"
      required:
        - type
        - event
    ItemHistoryEntry:
      type: object
      properties:
        date:
          type: string
          format: date-time
          nullable: true
        type:
          $ref: "#/components/schemas/ItemEventType"
        event:
          type: string
    ItemHistoryEntryListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ItemHistoryEntry"
        nextToken:
          type: string
          description: "Pagination token for next page"

    # Search
...
paths:
...
  /libraries/{libraryId}/items:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string

    get:
      summary: List library items
      description: Get paginated list of items in a library
      operationId: listLibraryItems
      tags:
        - Items
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 20
          description: "Number of items per page (max 20)"
        - name: nextToken
          in: query
          schema:
            type: string
          description: "Pagination token from previous response"
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetLibrariesContentResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /libraries/{libraryId}/books:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Create book
      description: Add a new book to a library
      operationId: createBook
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "200":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBookResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /libraries/{libraryId}/books/{bookId}:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
      - name: bookId
        in: path
        required: true
        schema:
          type: string

    put:
      summary: Update book
      description: Update book details
      operationId: updateBook
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: Book updated
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /libraries/{libraryId}/videos:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Create video
      description: Add a new video (DVD/Bluray) to a library
      operationId: createVideo
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoRequest"
      responses:
        "200":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateVideoResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /libraries/{libraryId}/videos/{videoId}:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
      - name: videoId
        in: path
        required: true
        schema:
          type: string

    put:
      summary: Update video
      description: Update video details
      operationId: updateVideo
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoRequest"
      responses:
        "200":
          description: Video updated
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /libraries/{libraryId}/items/{itemId}:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        required: true
        schema:
          type: string

    delete:
      summary: Delete item
      description: Delete an item from a library
      operationId: deleteItem
      tags:
        - Items
      responses:
        "200":
          description: Item deleted
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /libraries/{libraryId}/items/{itemId}/events:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get item history
      description: Get paginated lending/return history for an item
      operationId: getItemHistoryEvents
      tags:
        - Item History
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 20
          description: "Number of events per page (max 20)"
        - name: nextToken
          in: query
          schema:
            type: string
          description: "Pagination token from previous response"
      responses:
        "200":
          description: Item history events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemHistoryEntryListResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create history event
      description: Record a lend or return event for an item
      operationId: createItemHistoryEvent
      tags:
        - Item History
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemHistoryEntryRequest"
      responses:
        "201":
          description: Event created
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete item history
      description: Delete all history events for an item
      operationId: deleteItemHistoryEvents
      tags:
        - Item History
      responses:
        "200":
          description: History deleted
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...